refactor(ast): use nested struct definition for context to reduce the maintain cost
