+++ pilota-build/src/middle/resolver.rs
@@ -121,11 +121,13 @@ impl PathResolver for WorkspacePathResolver {
                 path.extend(prefix.iter().cloned());
                 path
             }
-            Some(super::context::DefLocation::Dynamic) => [cx.common_crate_name.clone().into()]
-                .iter()
-                .chain(DefaultPathResolver.mod_prefix(cx, def_id).iter())
-                .cloned()
-                .collect_vec(),
+            Some(super::context::DefLocation::Dynamic) => {
+                [cx.config.common_crate_name.clone().into()]
+                    .iter()
+                    .chain(DefaultPathResolver.mod_prefix(cx, def_id).iter())
+                    .cloned()
+                    .collect_vec()
+            }
             None => {
                 panic!(
                     "no location found for \"{}\" in file \"{}\"",
