+++ pilota-build/test_data/protobuf/nested_message.rs
@@ -1,5 +1,6 @@
 pub mod nested_message {
     #![allow(warnings, clippy::all)]
+    use ::pilota::{Buf as _, BufMut as _};
     #[derive(Debug, Default, Clone, PartialEq)]
     pub struct Tt1 {
         pub t2: tt1::T2,
@@ -8,84 +9,63 @@ pub mod nested_message {
 
         pub t4: tt1::t2::Tt3,
     }
-    impl ::pilota::prost::Message for Tt1 {
+    impl ::pilota::pb::Message for Tt1 {
         #[inline]
         fn encoded_len(&self) -> usize {
-            0 + ::pilota::prost::encoding::message::encoded_len(1, &self.t2)
-                + ::pilota::prost::encoding::int32::encoded_len(2, &self.t3)
-                + ::pilota::prost::encoding::message::encoded_len(4, &self.t4)
+            0 + ::pilota::pb::encoding::message::encoded_len(1, &self.t2)
+                + ::pilota::pb::encoding::int32::encoded_len(2, &self.t3)
+                + ::pilota::pb::encoding::message::encoded_len(4, &self.t4)
         }
 
         #[allow(unused_variables)]
-        fn encode_raw<B>(&self, buf: &mut B)
-        where
-            B: ::pilota::prost::bytes::BufMut,
-        {
-            ::pilota::prost::encoding::message::encode(1, (&self.t2), buf);
-            ::pilota::prost::encoding::int32::encode(2, &self.t3, buf);
-            ::pilota::prost::encoding::message::encode(4, (&self.t4), buf);
+        fn encode_raw(&self, buf: &mut ::pilota::LinkedBytes) {
+            ::pilota::pb::encoding::message::encode(1, (&self.t2), buf);
+            ::pilota::pb::encoding::int32::encode(2, &self.t3, buf);
+            ::pilota::pb::encoding::message::encode(4, (&self.t4), buf);
         }
 
         #[allow(unused_variables)]
-        fn merge_field<B>(
+        fn merge_field(
             &mut self,
             tag: u32,
-            wire_type: ::pilota::prost::encoding::WireType,
-            buf: &mut B,
-            ctx: ::pilota::prost::encoding::DecodeContext,
-        ) -> ::core::result::Result<(), ::pilota::prost::DecodeError>
-        where
-            B: ::pilota::prost::bytes::Buf,
-        {
+            wire_type: ::pilota::pb::encoding::WireType,
+            buf: &mut ::pilota::Bytes,
+            ctx: &mut ::pilota::pb::encoding::DecodeContext,
+        ) -> ::core::result::Result<(), ::pilota::pb::DecodeError> {
             const STRUCT_NAME: &'static str = stringify!(Tt1);
 
             match tag {
                 1 => {
                     let mut _inner_pilota_value = &mut self.t2;
-                    ::pilota::prost::encoding::message::merge(
-                        wire_type,
-                        _inner_pilota_value,
-                        buf,
-                        ctx,
-                    )
-                    .map_err(|mut error| {
-                        error.push(STRUCT_NAME, stringify!(t2));
-                        error
-                    })
+                    ::pilota::pb::encoding::message::merge(wire_type, _inner_pilota_value, buf, ctx)
+                        .map_err(|mut error| {
+                            error.push(STRUCT_NAME, stringify!(t2));
+                            error
+                        })
                 }
                 2 => {
                     let mut _inner_pilota_value = &mut self.t3;
-                    ::pilota::prost::encoding::int32::merge(
-                        wire_type,
-                        _inner_pilota_value,
-                        buf,
-                        ctx,
-                    )
-                    .map_err(|mut error| {
-                        error.push(STRUCT_NAME, stringify!(t3));
-                        error
-                    })
+                    ::pilota::pb::encoding::int32::merge(wire_type, _inner_pilota_value, buf, ctx)
+                        .map_err(|mut error| {
+                            error.push(STRUCT_NAME, stringify!(t3));
+                            error
+                        })
                 }
                 4 => {
                     let mut _inner_pilota_value = &mut self.t4;
-                    ::pilota::prost::encoding::message::merge(
-                        wire_type,
-                        _inner_pilota_value,
-                        buf,
-                        ctx,
-                    )
-                    .map_err(|mut error| {
-                        error.push(STRUCT_NAME, stringify!(t4));
-                        error
-                    })
+                    ::pilota::pb::encoding::message::merge(wire_type, _inner_pilota_value, buf, ctx)
+                        .map_err(|mut error| {
+                            error.push(STRUCT_NAME, stringify!(t4));
+                            error
+                        })
                 }
-                _ => ::pilota::prost::encoding::skip_field(wire_type, tag, buf, ctx),
+                _ => ::pilota::pb::encoding::skip_field(wire_type, tag, buf, ctx),
             }
         }
     }
 
     pub mod tt1 {
-
+        use ::pilota::{Buf as _, BufMut as _};
         #[derive(PartialOrd, Hash, Eq, Ord, Debug, Default, Clone, PartialEq, Copy)]
         #[repr(transparent)]
         pub struct Label(i32);
@@ -134,37 +114,31 @@ pub mod nested_message {
         pub struct T2 {
             pub t3: t2::Tt3,
         }
-        impl ::pilota::prost::Message for T2 {
+        impl ::pilota::pb::Message for T2 {
             #[inline]
             fn encoded_len(&self) -> usize {
-                0 + ::pilota::prost::encoding::message::encoded_len(1, &self.t3)
+                0 + ::pilota::pb::encoding::message::encoded_len(1, &self.t3)
             }
 
             #[allow(unused_variables)]
-            fn encode_raw<B>(&self, buf: &mut B)
-            where
-                B: ::pilota::prost::bytes::BufMut,
-            {
-                ::pilota::prost::encoding::message::encode(1, (&self.t3), buf);
+            fn encode_raw(&self, buf: &mut ::pilota::LinkedBytes) {
+                ::pilota::pb::encoding::message::encode(1, (&self.t3), buf);
             }
 
             #[allow(unused_variables)]
-            fn merge_field<B>(
+            fn merge_field(
                 &mut self,
                 tag: u32,
-                wire_type: ::pilota::prost::encoding::WireType,
-                buf: &mut B,
-                ctx: ::pilota::prost::encoding::DecodeContext,
-            ) -> ::core::result::Result<(), ::pilota::prost::DecodeError>
-            where
-                B: ::pilota::prost::bytes::Buf,
-            {
+                wire_type: ::pilota::pb::encoding::WireType,
+                buf: &mut ::pilota::Bytes,
+                ctx: &mut ::pilota::pb::encoding::DecodeContext,
+            ) -> ::core::result::Result<(), ::pilota::pb::DecodeError> {
                 const STRUCT_NAME: &'static str = stringify!(T2);
 
                 match tag {
                     1 => {
                         let mut _inner_pilota_value = &mut self.t3;
-                        ::pilota::prost::encoding::message::merge(
+                        ::pilota::pb::encoding::message::merge(
                             wire_type,
                             _inner_pilota_value,
                             buf,
@@ -175,45 +149,42 @@ pub mod nested_message {
                             error
                         })
                     }
-                    _ => ::pilota::prost::encoding::skip_field(wire_type, tag, buf, ctx),
+                    _ => ::pilota::pb::encoding::skip_field(wire_type, tag, buf, ctx),
                 }
             }
         }
 
         pub mod t2 {
-
+            use ::pilota::{Buf as _, BufMut as _};
             #[derive(Debug, Default, Clone, PartialEq)]
             pub struct Tt3 {
                 pub a: ::std::option::Option<i32>,
 
                 pub m: ::pilota::AHashMap<i32, super::T2>,
             }
-            impl ::pilota::prost::Message for Tt3 {
+            impl ::pilota::pb::Message for Tt3 {
                 #[inline]
                 fn encoded_len(&self) -> usize {
                     0 + self.a.as_ref().map_or(0, |value| {
-                        ::pilota::prost::encoding::int32::encoded_len(1, value)
-                    }) + ::pilota::prost::encoding::hash_map::encoded_len(
-                        ::pilota::prost::encoding::int32::encoded_len,
-                        ::pilota::prost::encoding::message::encoded_len,
+                        ::pilota::pb::encoding::int32::encoded_len(1, value)
+                    }) + ::pilota::pb::encoding::hash_map::encoded_len(
+                        ::pilota::pb::encoding::int32::encoded_len,
+                        ::pilota::pb::encoding::message::encoded_len,
                         2,
                         &self.m,
                     )
                 }
 
                 #[allow(unused_variables)]
-                fn encode_raw<B>(&self, buf: &mut B)
-                where
-                    B: ::pilota::prost::bytes::BufMut,
-                {
+                fn encode_raw(&self, buf: &mut ::pilota::LinkedBytes) {
                     if let Some(_pilota_inner_value) = self.a.as_ref() {
-                        ::pilota::prost::encoding::int32::encode(1, _pilota_inner_value, buf);
+                        ::pilota::pb::encoding::int32::encode(1, _pilota_inner_value, buf);
                     };
-                    ::pilota::prost::encoding::hash_map::encode(
-                        ::pilota::prost::encoding::int32::encode,
-                        ::pilota::prost::encoding::int32::encoded_len,
-                        ::pilota::prost::encoding::message::encode,
-                        ::pilota::prost::encoding::message::encoded_len,
+                    ::pilota::pb::encoding::hash_map::encode(
+                        ::pilota::pb::encoding::int32::encode,
+                        ::pilota::pb::encoding::int32::encoded_len,
+                        ::pilota::pb::encoding::message::encode,
+                        ::pilota::pb::encoding::message::encoded_len,
                         2,
                         &self.m,
                         buf,
@@ -221,22 +192,19 @@ pub mod nested_message {
                 }
 
                 #[allow(unused_variables)]
-                fn merge_field<B>(
+                fn merge_field(
                     &mut self,
                     tag: u32,
-                    wire_type: ::pilota::prost::encoding::WireType,
-                    buf: &mut B,
-                    ctx: ::pilota::prost::encoding::DecodeContext,
-                ) -> ::core::result::Result<(), ::pilota::prost::DecodeError>
-                where
-                    B: ::pilota::prost::bytes::Buf,
-                {
+                    wire_type: ::pilota::pb::encoding::WireType,
+                    buf: &mut ::pilota::Bytes,
+                    ctx: &mut ::pilota::pb::encoding::DecodeContext,
+                ) -> ::core::result::Result<(), ::pilota::pb::DecodeError> {
                     const STRUCT_NAME: &'static str = stringify!(Tt3);
 
                     match tag {
                         1 => {
                             let mut _inner_pilota_value = &mut self.a;
-                            ::pilota::prost::encoding::int32::merge(
+                            ::pilota::pb::encoding::int32::merge(
                                 wire_type,
                                 _inner_pilota_value
                                     .get_or_insert_with(::core::default::Default::default),
@@ -250,9 +218,9 @@ pub mod nested_message {
                         }
                         2 => {
                             let mut _inner_pilota_value = &mut self.m;
-                            ::pilota::prost::encoding::hash_map::merge(
-                                ::pilota::prost::encoding::int32::merge,
-                                ::pilota::prost::encoding::message::merge,
+                            ::pilota::pb::encoding::hash_map::merge(
+                                ::pilota::pb::encoding::int32::merge,
+                                ::pilota::pb::encoding::message::merge,
                                 &mut _inner_pilota_value,
                                 buf,
                                 ctx,
@@ -262,7 +230,7 @@ pub mod nested_message {
                                 error
                             })
                         }
-                        _ => ::pilota::prost::encoding::skip_field(wire_type, tag, buf, ctx),
+                        _ => ::pilota::pb::encoding::skip_field(wire_type, tag, buf, ctx),
                     }
                 }
             }
