+++ pilota-build/test_data/protobuf/normal.rs
@@ -1,51 +1,41 @@
 pub mod normal {
     #![allow(warnings, clippy::all)]
+    use ::pilota::{Buf as _, BufMut as _};
     #[derive(PartialOrd, Hash, Eq, Ord, Debug, Default, Clone, PartialEq)]
     pub struct A {
         pub a: i32,
     }
-    impl ::pilota::prost::Message for A {
+    impl ::pilota::pb::Message for A {
         #[inline]
         fn encoded_len(&self) -> usize {
-            0 + ::pilota::prost::encoding::int32::encoded_len(1, &self.a)
+            0 + ::pilota::pb::encoding::int32::encoded_len(1, &self.a)
         }
 
         #[allow(unused_variables)]
-        fn encode_raw<B>(&self, buf: &mut B)
-        where
-            B: ::pilota::prost::bytes::BufMut,
-        {
-            ::pilota::prost::encoding::int32::encode(1, &self.a, buf);
+        fn encode_raw(&self, buf: &mut ::pilota::LinkedBytes) {
+            ::pilota::pb::encoding::int32::encode(1, &self.a, buf);
         }
 
         #[allow(unused_variables)]
-        fn merge_field<B>(
+        fn merge_field(
             &mut self,
             tag: u32,
-            wire_type: ::pilota::prost::encoding::WireType,
-            buf: &mut B,
-            ctx: ::pilota::prost::encoding::DecodeContext,
-        ) -> ::core::result::Result<(), ::pilota::prost::DecodeError>
-        where
-            B: ::pilota::prost::bytes::Buf,
-        {
+            wire_type: ::pilota::pb::encoding::WireType,
+            buf: &mut ::pilota::Bytes,
+            ctx: &mut ::pilota::pb::encoding::DecodeContext,
+        ) -> ::core::result::Result<(), ::pilota::pb::DecodeError> {
             const STRUCT_NAME: &'static str = stringify!(A);
 
             match tag {
                 1 => {
                     let mut _inner_pilota_value = &mut self.a;
-                    ::pilota::prost::encoding::int32::merge(
-                        wire_type,
-                        _inner_pilota_value,
-                        buf,
-                        ctx,
-                    )
-                    .map_err(|mut error| {
-                        error.push(STRUCT_NAME, stringify!(a));
-                        error
-                    })
+                    ::pilota::pb::encoding::int32::merge(wire_type, _inner_pilota_value, buf, ctx)
+                        .map_err(|mut error| {
+                            error.push(STRUCT_NAME, stringify!(a));
+                            error
+                        })
                 }
-                _ => ::pilota::prost::encoding::skip_field(wire_type, tag, buf, ctx),
+                _ => ::pilota::pb::encoding::skip_field(wire_type, tag, buf, ctx),
             }
         }
     }
@@ -53,41 +43,35 @@ pub mod normal {
     pub struct SubMessage {
         pub value: ::std::option::Option<::pilota::FastStr>,
     }
-    impl ::pilota::prost::Message for SubMessage {
+    impl ::pilota::pb::Message for SubMessage {
         #[inline]
         fn encoded_len(&self) -> usize {
             0 + self.value.as_ref().map_or(0, |value| {
-                ::pilota::prost::encoding::faststr::encoded_len(2, value)
+                ::pilota::pb::encoding::faststr::encoded_len(2, value)
             })
         }
 
         #[allow(unused_variables)]
-        fn encode_raw<B>(&self, buf: &mut B)
-        where
-            B: ::pilota::prost::bytes::BufMut,
-        {
+        fn encode_raw(&self, buf: &mut ::pilota::LinkedBytes) {
             if let Some(_pilota_inner_value) = self.value.as_ref() {
-                ::pilota::prost::encoding::faststr::encode(2, _pilota_inner_value, buf);
+                ::pilota::pb::encoding::faststr::encode(2, _pilota_inner_value, buf);
             };
         }
 
         #[allow(unused_variables)]
-        fn merge_field<B>(
+        fn merge_field(
             &mut self,
             tag: u32,
-            wire_type: ::pilota::prost::encoding::WireType,
-            buf: &mut B,
-            ctx: ::pilota::prost::encoding::DecodeContext,
-        ) -> ::core::result::Result<(), ::pilota::prost::DecodeError>
-        where
-            B: ::pilota::prost::bytes::Buf,
-        {
+            wire_type: ::pilota::pb::encoding::WireType,
+            buf: &mut ::pilota::Bytes,
+            ctx: &mut ::pilota::pb::encoding::DecodeContext,
+        ) -> ::core::result::Result<(), ::pilota::pb::DecodeError> {
             const STRUCT_NAME: &'static str = stringify!(SubMessage);
 
             match tag {
                 2 => {
                     let mut _inner_pilota_value = &mut self.value;
-                    ::pilota::prost::encoding::faststr::merge(
+                    ::pilota::pb::encoding::faststr::merge(
                         wire_type,
                         _inner_pilota_value.get_or_insert_with(::core::default::Default::default),
                         buf,
@@ -98,7 +82,7 @@ pub mod normal {
                         error
                     })
                 }
-                _ => ::pilota::prost::encoding::skip_field(wire_type, tag, buf, ctx),
+                _ => ::pilota::pb::encoding::skip_field(wire_type, tag, buf, ctx),
             }
         }
     }
@@ -118,60 +102,54 @@ pub mod normal {
 
         pub mock_cost: ::std::option::Option<::pilota::FastStr>,
     }
-    impl ::pilota::prost::Message for ObjReq {
+    impl ::pilota::pb::Message for ObjReq {
         #[inline]
         fn encoded_len(&self) -> usize {
             0 + self.msg.as_ref().map_or(0, |msg| {
-                ::pilota::prost::encoding::message::encoded_len(1, msg)
-            }) + ::pilota::prost::encoding::message::encoded_len_repeated(2, &self.msg_map)
-                + ::pilota::prost::encoding::message::encoded_len_repeated(3, &self.sub_msgs)
-                + ::pilota::prost::encoding::message::encoded_len_repeated(4, &self.msg_set)
-                + ::pilota::prost::encoding::faststr::encoded_len(5, &self.flag_msg)
+                ::pilota::pb::encoding::message::encoded_len(1, msg)
+            }) + ::pilota::pb::encoding::message::encoded_len_repeated(2, &self.msg_map)
+                + ::pilota::pb::encoding::message::encoded_len_repeated(3, &self.sub_msgs)
+                + ::pilota::pb::encoding::message::encoded_len_repeated(4, &self.msg_set)
+                + ::pilota::pb::encoding::faststr::encoded_len(5, &self.flag_msg)
                 + self.mock_cost.as_ref().map_or(0, |value| {
-                    ::pilota::prost::encoding::faststr::encoded_len(6, value)
+                    ::pilota::pb::encoding::faststr::encoded_len(6, value)
                 })
         }
 
         #[allow(unused_variables)]
-        fn encode_raw<B>(&self, buf: &mut B)
-        where
-            B: ::pilota::prost::bytes::BufMut,
-        {
+        fn encode_raw(&self, buf: &mut ::pilota::LinkedBytes) {
             if let Some(_pilota_inner_value) = self.msg.as_ref() {
-                ::pilota::prost::encoding::message::encode(1, _pilota_inner_value, buf);
+                ::pilota::pb::encoding::message::encode(1, _pilota_inner_value, buf);
             }
             for msg in &self.msg_map {
-                ::pilota::prost::encoding::message::encode(2, msg, buf);
+                ::pilota::pb::encoding::message::encode(2, msg, buf);
             }
             for msg in &self.sub_msgs {
-                ::pilota::prost::encoding::message::encode(3, msg, buf);
+                ::pilota::pb::encoding::message::encode(3, msg, buf);
             }
             for msg in &self.msg_set {
-                ::pilota::prost::encoding::message::encode(4, msg, buf);
+                ::pilota::pb::encoding::message::encode(4, msg, buf);
             }
-            ::pilota::prost::encoding::faststr::encode(5, &self.flag_msg, buf);
+            ::pilota::pb::encoding::faststr::encode(5, &self.flag_msg, buf);
             if let Some(_pilota_inner_value) = self.mock_cost.as_ref() {
-                ::pilota::prost::encoding::faststr::encode(6, _pilota_inner_value, buf);
+                ::pilota::pb::encoding::faststr::encode(6, _pilota_inner_value, buf);
             };
         }
 
         #[allow(unused_variables)]
-        fn merge_field<B>(
+        fn merge_field(
             &mut self,
             tag: u32,
-            wire_type: ::pilota::prost::encoding::WireType,
-            buf: &mut B,
-            ctx: ::pilota::prost::encoding::DecodeContext,
-        ) -> ::core::result::Result<(), ::pilota::prost::DecodeError>
-        where
-            B: ::pilota::prost::bytes::Buf,
-        {
+            wire_type: ::pilota::pb::encoding::WireType,
+            buf: &mut ::pilota::Bytes,
+            ctx: &mut ::pilota::pb::encoding::DecodeContext,
+        ) -> ::core::result::Result<(), ::pilota::pb::DecodeError> {
             const STRUCT_NAME: &'static str = stringify!(ObjReq);
 
             match tag {
                 1 => {
                     let mut _inner_pilota_value = &mut self.msg;
-                    ::pilota::prost::encoding::message::merge(
+                    ::pilota::pb::encoding::message::merge(
                         wire_type,
                         _inner_pilota_value.get_or_insert_with(::core::default::Default::default),
                         buf,
@@ -184,7 +162,7 @@ pub mod normal {
                 }
                 2 => {
                     let mut _inner_pilota_value = &mut self.msg_map;
-                    ::pilota::prost::encoding::message::merge_repeated(
+                    ::pilota::pb::encoding::message::merge_repeated(
                         wire_type,
                         _inner_pilota_value,
                         buf,
@@ -197,7 +175,7 @@ pub mod normal {
                 }
                 3 => {
                     let mut _inner_pilota_value = &mut self.sub_msgs;
-                    ::pilota::prost::encoding::message::merge_repeated(
+                    ::pilota::pb::encoding::message::merge_repeated(
                         wire_type,
                         _inner_pilota_value,
                         buf,
@@ -210,7 +188,7 @@ pub mod normal {
                 }
                 4 => {
                     let mut _inner_pilota_value = &mut self.msg_set;
-                    ::pilota::prost::encoding::message::merge_repeated(
+                    ::pilota::pb::encoding::message::merge_repeated(
                         wire_type,
                         _inner_pilota_value,
                         buf,
@@ -223,20 +201,15 @@ pub mod normal {
                 }
                 5 => {
                     let mut _inner_pilota_value = &mut self.flag_msg;
-                    ::pilota::prost::encoding::faststr::merge(
-                        wire_type,
-                        _inner_pilota_value,
-                        buf,
-                        ctx,
-                    )
-                    .map_err(|mut error| {
-                        error.push(STRUCT_NAME, stringify!(flag_msg));
-                        error
-                    })
+                    ::pilota::pb::encoding::faststr::merge(wire_type, _inner_pilota_value, buf, ctx)
+                        .map_err(|mut error| {
+                            error.push(STRUCT_NAME, stringify!(flag_msg));
+                            error
+                        })
                 }
                 6 => {
                     let mut _inner_pilota_value = &mut self.mock_cost;
-                    ::pilota::prost::encoding::faststr::merge(
+                    ::pilota::pb::encoding::faststr::merge(
                         wire_type,
                         _inner_pilota_value.get_or_insert_with(::core::default::Default::default),
                         buf,
@@ -247,7 +220,7 @@ pub mod normal {
                         error
                     })
                 }
-                _ => ::pilota::prost::encoding::skip_field(wire_type, tag, buf, ctx),
+                _ => ::pilota::pb::encoding::skip_field(wire_type, tag, buf, ctx),
             }
         }
     }
@@ -255,41 +228,35 @@ pub mod normal {
     pub struct B {
         pub a: ::std::option::Option<A>,
     }
-    impl ::pilota::prost::Message for B {
+    impl ::pilota::pb::Message for B {
         #[inline]
         fn encoded_len(&self) -> usize {
             0 + self.a.as_ref().map_or(0, |msg| {
-                ::pilota::prost::encoding::message::encoded_len(2, msg)
+                ::pilota::pb::encoding::message::encoded_len(2, msg)
             })
         }
 
         #[allow(unused_variables)]
-        fn encode_raw<B>(&self, buf: &mut B)
-        where
-            B: ::pilota::prost::bytes::BufMut,
-        {
+        fn encode_raw(&self, buf: &mut ::pilota::LinkedBytes) {
             if let Some(_pilota_inner_value) = self.a.as_ref() {
-                ::pilota::prost::encoding::message::encode(2, _pilota_inner_value, buf);
+                ::pilota::pb::encoding::message::encode(2, _pilota_inner_value, buf);
             }
         }
 
         #[allow(unused_variables)]
-        fn merge_field<B>(
+        fn merge_field(
             &mut self,
             tag: u32,
-            wire_type: ::pilota::prost::encoding::WireType,
-            buf: &mut B,
-            ctx: ::pilota::prost::encoding::DecodeContext,
-        ) -> ::core::result::Result<(), ::pilota::prost::DecodeError>
-        where
-            B: ::pilota::prost::bytes::Buf,
-        {
+            wire_type: ::pilota::pb::encoding::WireType,
+            buf: &mut ::pilota::Bytes,
+            ctx: &mut ::pilota::pb::encoding::DecodeContext,
+        ) -> ::core::result::Result<(), ::pilota::pb::DecodeError> {
             const STRUCT_NAME: &'static str = stringify!(B);
 
             match tag {
                 2 => {
                     let mut _inner_pilota_value = &mut self.a;
-                    ::pilota::prost::encoding::message::merge(
+                    ::pilota::pb::encoding::message::merge(
                         wire_type,
                         _inner_pilota_value.get_or_insert_with(::core::default::Default::default),
                         buf,
@@ -300,7 +267,7 @@ pub mod normal {
                         error
                     })
                 }
-                _ => ::pilota::prost::encoding::skip_field(wire_type, tag, buf, ctx),
+                _ => ::pilota::pb::encoding::skip_field(wire_type, tag, buf, ctx),
             }
         }
     }
@@ -312,60 +279,49 @@ pub mod normal {
 
         pub sub_messages: ::std::vec::Vec<SubMessage>,
     }
-    impl ::pilota::prost::Message for Message {
+    impl ::pilota::pb::Message for Message {
         #[inline]
         fn encoded_len(&self) -> usize {
-            0 + ::pilota::prost::encoding::faststr::encoded_len(1, &self.uid)
+            0 + ::pilota::pb::encoding::faststr::encoded_len(1, &self.uid)
                 + self.value.as_ref().map_or(0, |value| {
-                    ::pilota::prost::encoding::faststr::encoded_len(2, value)
+                    ::pilota::pb::encoding::faststr::encoded_len(2, value)
                 })
-                + ::pilota::prost::encoding::message::encoded_len_repeated(3, &self.sub_messages)
+                + ::pilota::pb::encoding::message::encoded_len_repeated(3, &self.sub_messages)
         }
 
         #[allow(unused_variables)]
-        fn encode_raw<B>(&self, buf: &mut B)
-        where
-            B: ::pilota::prost::bytes::BufMut,
-        {
-            ::pilota::prost::encoding::faststr::encode(1, &self.uid, buf);
+        fn encode_raw(&self, buf: &mut ::pilota::LinkedBytes) {
+            ::pilota::pb::encoding::faststr::encode(1, &self.uid, buf);
             if let Some(_pilota_inner_value) = self.value.as_ref() {
-                ::pilota::prost::encoding::faststr::encode(2, _pilota_inner_value, buf);
+                ::pilota::pb::encoding::faststr::encode(2, _pilota_inner_value, buf);
             };
             for msg in &self.sub_messages {
-                ::pilota::prost::encoding::message::encode(3, msg, buf);
+                ::pilota::pb::encoding::message::encode(3, msg, buf);
             }
         }
 
         #[allow(unused_variables)]
-        fn merge_field<B>(
+        fn merge_field(
             &mut self,
             tag: u32,
-            wire_type: ::pilota::prost::encoding::WireType,
-            buf: &mut B,
-            ctx: ::pilota::prost::encoding::DecodeContext,
-        ) -> ::core::result::Result<(), ::pilota::prost::DecodeError>
-        where
-            B: ::pilota::prost::bytes::Buf,
-        {
+            wire_type: ::pilota::pb::encoding::WireType,
+            buf: &mut ::pilota::Bytes,
+            ctx: &mut ::pilota::pb::encoding::DecodeContext,
+        ) -> ::core::result::Result<(), ::pilota::pb::DecodeError> {
             const STRUCT_NAME: &'static str = stringify!(Message);
 
             match tag {
                 1 => {
                     let mut _inner_pilota_value = &mut self.uid;
-                    ::pilota::prost::encoding::faststr::merge(
-                        wire_type,
-                        _inner_pilota_value,
-                        buf,
-                        ctx,
-                    )
-                    .map_err(|mut error| {
-                        error.push(STRUCT_NAME, stringify!(uid));
-                        error
-                    })
+                    ::pilota::pb::encoding::faststr::merge(wire_type, _inner_pilota_value, buf, ctx)
+                        .map_err(|mut error| {
+                            error.push(STRUCT_NAME, stringify!(uid));
+                            error
+                        })
                 }
                 2 => {
                     let mut _inner_pilota_value = &mut self.value;
-                    ::pilota::prost::encoding::faststr::merge(
+                    ::pilota::pb::encoding::faststr::merge(
                         wire_type,
                         _inner_pilota_value.get_or_insert_with(::core::default::Default::default),
                         buf,
@@ -378,7 +334,7 @@ pub mod normal {
                 }
                 3 => {
                     let mut _inner_pilota_value = &mut self.sub_messages;
-                    ::pilota::prost::encoding::message::merge_repeated(
+                    ::pilota::pb::encoding::message::merge_repeated(
                         wire_type,
                         _inner_pilota_value,
                         buf,
@@ -389,59 +345,53 @@ pub mod normal {
                         error
                     })
                 }
-                _ => ::pilota::prost::encoding::skip_field(wire_type, tag, buf, ctx),
+                _ => ::pilota::pb::encoding::skip_field(wire_type, tag, buf, ctx),
             }
         }
     }
 
     pub mod obj_req {
-
+        use ::pilota::{Buf as _, BufMut as _};
         #[derive(PartialOrd, Hash, Eq, Ord, Debug, Default, Clone, PartialEq)]
         pub struct MsgMapEntry {
             pub key: ::std::option::Option<super::Message>,
 
             pub value: ::std::option::Option<super::SubMessage>,
         }
-        impl ::pilota::prost::Message for MsgMapEntry {
+        impl ::pilota::pb::Message for MsgMapEntry {
             #[inline]
             fn encoded_len(&self) -> usize {
                 0 + self.key.as_ref().map_or(0, |msg| {
-                    ::pilota::prost::encoding::message::encoded_len(1, msg)
+                    ::pilota::pb::encoding::message::encoded_len(1, msg)
                 }) + self.value.as_ref().map_or(0, |msg| {
-                    ::pilota::prost::encoding::message::encoded_len(2, msg)
+                    ::pilota::pb::encoding::message::encoded_len(2, msg)
                 })
             }
 
             #[allow(unused_variables)]
-            fn encode_raw<B>(&self, buf: &mut B)
-            where
-                B: ::pilota::prost::bytes::BufMut,
-            {
+            fn encode_raw(&self, buf: &mut ::pilota::LinkedBytes) {
                 if let Some(_pilota_inner_value) = self.key.as_ref() {
-                    ::pilota::prost::encoding::message::encode(1, _pilota_inner_value, buf);
+                    ::pilota::pb::encoding::message::encode(1, _pilota_inner_value, buf);
                 }
                 if let Some(_pilota_inner_value) = self.value.as_ref() {
-                    ::pilota::prost::encoding::message::encode(2, _pilota_inner_value, buf);
+                    ::pilota::pb::encoding::message::encode(2, _pilota_inner_value, buf);
                 }
             }
 
             #[allow(unused_variables)]
-            fn merge_field<B>(
+            fn merge_field(
                 &mut self,
                 tag: u32,
-                wire_type: ::pilota::prost::encoding::WireType,
-                buf: &mut B,
-                ctx: ::pilota::prost::encoding::DecodeContext,
-            ) -> ::core::result::Result<(), ::pilota::prost::DecodeError>
-            where
-                B: ::pilota::prost::bytes::Buf,
-            {
+                wire_type: ::pilota::pb::encoding::WireType,
+                buf: &mut ::pilota::Bytes,
+                ctx: &mut ::pilota::pb::encoding::DecodeContext,
+            ) -> ::core::result::Result<(), ::pilota::pb::DecodeError> {
                 const STRUCT_NAME: &'static str = stringify!(MsgMapEntry);
 
                 match tag {
                     1 => {
                         let mut _inner_pilota_value = &mut self.key;
-                        ::pilota::prost::encoding::message::merge(
+                        ::pilota::pb::encoding::message::merge(
                             wire_type,
                             _inner_pilota_value
                                 .get_or_insert_with(::core::default::Default::default),
@@ -455,7 +405,7 @@ pub mod normal {
                     }
                     2 => {
                         let mut _inner_pilota_value = &mut self.value;
-                        ::pilota::prost::encoding::message::merge(
+                        ::pilota::pb::encoding::message::merge(
                             wire_type,
                             _inner_pilota_value
                                 .get_or_insert_with(::core::default::Default::default),
@@ -467,7 +417,7 @@ pub mod normal {
                             error
                         })
                     }
-                    _ => ::pilota::prost::encoding::skip_field(wire_type, tag, buf, ctx),
+                    _ => ::pilota::pb::encoding::skip_field(wire_type, tag, buf, ctx),
                 }
             }
         }
