+++ pilota-build/test_data/protobuf/string.rs
@@ -1,47 +1,42 @@
 pub mod string {
     #![allow(warnings, clippy::all)]
+    use ::pilota::{Buf as _, BufMut as _};
     #[derive(PartialOrd, Hash, Eq, Ord, Debug, Default, Clone, PartialEq)]
     pub struct A {
         pub a: ::std::option::Option<::pilota::FastStr>,
 
         pub b: ::pilota::FastStr,
     }
-    impl ::pilota::prost::Message for A {
+    impl ::pilota::pb::Message for A {
         #[inline]
         fn encoded_len(&self) -> usize {
             0 + self.a.as_ref().map_or(0, |value| {
-                ::pilota::prost::encoding::faststr::encoded_len(1, value)
-            }) + ::pilota::prost::encoding::faststr::encoded_len(2, &self.b)
+                ::pilota::pb::encoding::faststr::encoded_len(1, value)
+            }) + ::pilota::pb::encoding::faststr::encoded_len(2, &self.b)
         }
 
         #[allow(unused_variables)]
-        fn encode_raw<B>(&self, buf: &mut B)
-        where
-            B: ::pilota::prost::bytes::BufMut,
-        {
+        fn encode_raw(&self, buf: &mut ::pilota::LinkedBytes) {
             if let Some(_pilota_inner_value) = self.a.as_ref() {
-                ::pilota::prost::encoding::faststr::encode(1, _pilota_inner_value, buf);
+                ::pilota::pb::encoding::faststr::encode(1, _pilota_inner_value, buf);
             };
-            ::pilota::prost::encoding::faststr::encode(2, &self.b, buf);
+            ::pilota::pb::encoding::faststr::encode(2, &self.b, buf);
         }
 
         #[allow(unused_variables)]
-        fn merge_field<B>(
+        fn merge_field(
             &mut self,
             tag: u32,
-            wire_type: ::pilota::prost::encoding::WireType,
-            buf: &mut B,
-            ctx: ::pilota::prost::encoding::DecodeContext,
-        ) -> ::core::result::Result<(), ::pilota::prost::DecodeError>
-        where
-            B: ::pilota::prost::bytes::Buf,
-        {
+            wire_type: ::pilota::pb::encoding::WireType,
+            buf: &mut ::pilota::Bytes,
+            ctx: &mut ::pilota::pb::encoding::DecodeContext,
+        ) -> ::core::result::Result<(), ::pilota::pb::DecodeError> {
             const STRUCT_NAME: &'static str = stringify!(A);
 
             match tag {
                 1 => {
                     let mut _inner_pilota_value = &mut self.a;
-                    ::pilota::prost::encoding::faststr::merge(
+                    ::pilota::pb::encoding::faststr::merge(
                         wire_type,
                         _inner_pilota_value.get_or_insert_with(::core::default::Default::default),
                         buf,
@@ -54,18 +49,13 @@ pub mod string {
                 }
                 2 => {
                     let mut _inner_pilota_value = &mut self.b;
-                    ::pilota::prost::encoding::faststr::merge(
-                        wire_type,
-                        _inner_pilota_value,
-                        buf,
-                        ctx,
-                    )
-                    .map_err(|mut error| {
-                        error.push(STRUCT_NAME, stringify!(b));
-                        error
-                    })
+                    ::pilota::pb::encoding::faststr::merge(wire_type, _inner_pilota_value, buf, ctx)
+                        .map_err(|mut error| {
+                            error.push(STRUCT_NAME, stringify!(b));
+                            error
+                        })
                 }
-                _ => ::pilota::prost::encoding::skip_field(wire_type, tag, buf, ctx),
+                _ => ::pilota::pb::encoding::skip_field(wire_type, tag, buf, ctx),
             }
         }
     }
