--- pilota-build/test_data/protobuf_with_split/new_pb/service.rs
@@ -1,8 +1,0 @@
-pub mod service {
-    #![allow(warnings, clippy::all)]
-    use ::pilota::{Buf as _, BufMut as _};
-    pub mod service {
-        use ::pilota::{Buf as _, BufMut as _};
-        include!("service/mod.rs");
-    }
-}
