--- pilota-build/test_data/protobuf_with_split/new_pb/service/mod.rs
@@ -1,3 +1,0 @@
-include!("message_EchoRequest.rs");
-include!("service_Echo.rs");
-include!("message_EchoResponse.rs");
