--- pilota-build/test_data/protobuf_with_split/new_pb/service/service_Echo.rs
@@ -1,2 +1,0 @@
-
-pub trait Echo {}
