+++ pilota-build/test_data/protobuf_with_split/pb/service/message_EchoResponse.rs
@@ -2,43 +2,37 @@
 pub struct EchoResponse {
     pub message: ::pilota::FastStr,
 }
-impl ::pilota::prost::Message for EchoResponse {
+impl ::pilota::pb::Message for EchoResponse {
     #[inline]
     fn encoded_len(&self) -> usize {
-        0 + ::pilota::prost::encoding::faststr::encoded_len(1, &self.message)
+        0 + ::pilota::pb::encoding::faststr::encoded_len(1, &self.message)
     }
 
     #[allow(unused_variables)]
-    fn encode_raw<B>(&self, buf: &mut B)
-    where
-        B: ::pilota::prost::bytes::BufMut,
-    {
-        ::pilota::prost::encoding::faststr::encode(1, &self.message, buf);
+    fn encode_raw(&self, buf: &mut ::pilota::LinkedBytes) {
+        ::pilota::pb::encoding::faststr::encode(1, &self.message, buf);
     }
 
     #[allow(unused_variables)]
-    fn merge_field<B>(
+    fn merge_field(
         &mut self,
         tag: u32,
-        wire_type: ::pilota::prost::encoding::WireType,
-        buf: &mut B,
-        ctx: ::pilota::prost::encoding::DecodeContext,
-    ) -> ::core::result::Result<(), ::pilota::prost::DecodeError>
-    where
-        B: ::pilota::prost::bytes::Buf,
-    {
+        wire_type: ::pilota::pb::encoding::WireType,
+        buf: &mut ::pilota::Bytes,
+        ctx: &mut ::pilota::pb::encoding::DecodeContext,
+    ) -> ::core::result::Result<(), ::pilota::pb::DecodeError> {
         const STRUCT_NAME: &'static str = stringify!(EchoResponse);
 
         match tag {
             1 => {
                 let mut _inner_pilota_value = &mut self.message;
-                ::pilota::prost::encoding::faststr::merge(wire_type, _inner_pilota_value, buf, ctx)
+                ::pilota::pb::encoding::faststr::merge(wire_type, _inner_pilota_value, buf, ctx)
                     .map_err(|mut error| {
                         error.push(STRUCT_NAME, stringify!(message));
                         error
                     })
             }
-            _ => ::pilota::prost::encoding::skip_field(wire_type, tag, buf, ctx),
+            _ => ::pilota::pb::encoding::skip_field(wire_type, tag, buf, ctx),
         }
     }
 }
