+++ pilota-build/test_data/unknown_fields_pb.rs
@@ -1,5 +1,6 @@
 pub mod unknown_fields_pb {
     #![allow(warnings, clippy::all)]
+    use ::pilota::{Buf as _, BufMut as _};
     #[derive(
         PartialOrd,
         Hash,
@@ -16,61 +17,46 @@ pub mod unknown_fields_pb {
         pub a: i32,
         pub _unknown_fields: ::pilota::BytesVec,
     }
-    impl ::pilota::prost::Message for A {
+    impl ::pilota::pb::Message for A {
         #[inline]
         fn encoded_len(&self) -> usize {
-            0 + ::pilota::prost::encoding::int32::encoded_len(1, &self.a)
-                + self._unknown_fields.size()
+            0 + ::pilota::pb::encoding::int32::encoded_len(1, &self.a) + self._unknown_fields.size()
         }
 
         #[allow(unused_variables)]
-        fn encode_raw<B>(&self, buf: &mut B)
-        where
-            B: ::pilota::prost::bytes::BufMut,
-        {
-            ::pilota::prost::encoding::int32::encode(1, &self.a, buf);
+        fn encode_raw(&self, buf: &mut ::pilota::LinkedBytes) {
+            ::pilota::pb::encoding::int32::encode(1, &self.a, buf);
             for bytes in self._unknown_fields.list.iter() {
                 buf.put_slice(bytes.as_ref());
             }
         }
 
         #[allow(unused_variables)]
-        fn merge_field<B>(
+        fn merge_field(
             &mut self,
             tag: u32,
-            wire_type: ::pilota::prost::encoding::WireType,
-            buf: &mut B,
-            ctx: ::pilota::prost::encoding::DecodeContext,
-        ) -> ::core::result::Result<(), ::pilota::prost::DecodeError>
-        where
-            B: ::pilota::prost::bytes::Buf,
-        {
+            wire_type: ::pilota::pb::encoding::WireType,
+            buf: &mut ::pilota::Bytes,
+            ctx: &mut ::pilota::pb::encoding::DecodeContext,
+        ) -> ::core::result::Result<(), ::pilota::pb::DecodeError> {
             const STRUCT_NAME: &'static str = stringify!(A);
 
             let mut _unknown_fields = &mut self._unknown_fields;
             match tag {
                 1 => {
                     let mut _inner_pilota_value = &mut self.a;
-                    ::pilota::prost::encoding::int32::merge(
-                        wire_type,
-                        _inner_pilota_value,
-                        buf,
-                        ctx,
-                    )
-                    .map_err(|mut error| {
-                        error.push(STRUCT_NAME, stringify!(a));
-                        error
-                    })
+                    ::pilota::pb::encoding::int32::merge(wire_type, _inner_pilota_value, buf, ctx)
+                        .map_err(|mut error| {
+                            error.push(STRUCT_NAME, stringify!(a));
+                            error
+                        })
                 }
                 _ => {
-                    let tag_value = (tag << 3) | wire_type as u32;
-                    let tag_len = ::pilota::prost::encoding::encoded_len_varint(tag_value as u64);
-                    let begin = unsafe { buf.chunk().as_ptr().sub(tag_len) };
-                    ::pilota::prost::encoding::skip_field(wire_type, tag, buf, ctx)?;
+                    ::pilota::pb::encoding::skip_field(wire_type, tag, buf, ctx)?;
                     let end = buf.chunk().as_ptr();
-                    _unknown_fields.push_back(::pilota::Bytes::copy_from_slice(unsafe {
-                        std::slice::from_raw_parts(begin, end as usize - begin as usize)
-                    }));
+                    let len = end as usize - ctx.raw_bytes_cursor();
+                    let val = ctx.raw_bytes_split_to(len);
+                    _unknown_fields.push_back(val);
                     Ok(())
                 }
             }
@@ -91,44 +77,35 @@ pub mod unknown_fields_pb {
     pub struct ObjReq {
         pub _unknown_fields: ::pilota::BytesVec,
     }
-    impl ::pilota::prost::Message for ObjReq {
+    impl ::pilota::pb::Message for ObjReq {
         #[inline]
         fn encoded_len(&self) -> usize {
             0 + self._unknown_fields.size()
         }
 
         #[allow(unused_variables)]
-        fn encode_raw<B>(&self, buf: &mut B)
-        where
-            B: ::pilota::prost::bytes::BufMut,
-        {
+        fn encode_raw(&self, buf: &mut ::pilota::LinkedBytes) {
             for bytes in self._unknown_fields.list.iter() {
                 buf.put_slice(bytes.as_ref());
             }
         }
 
         #[allow(unused_variables)]
-        fn merge_field<B>(
+        fn merge_field(
             &mut self,
             tag: u32,
-            wire_type: ::pilota::prost::encoding::WireType,
-            buf: &mut B,
-            ctx: ::pilota::prost::encoding::DecodeContext,
-        ) -> ::core::result::Result<(), ::pilota::prost::DecodeError>
-        where
-            B: ::pilota::prost::bytes::Buf,
-        {
+            wire_type: ::pilota::pb::encoding::WireType,
+            buf: &mut ::pilota::Bytes,
+            ctx: &mut ::pilota::pb::encoding::DecodeContext,
+        ) -> ::core::result::Result<(), ::pilota::pb::DecodeError> {
             let mut _unknown_fields = &mut self._unknown_fields;
             match tag {
                 _ => {
-                    let tag_value = (tag << 3) | wire_type as u32;
-                    let tag_len = ::pilota::prost::encoding::encoded_len_varint(tag_value as u64);
-                    let begin = unsafe { buf.chunk().as_ptr().sub(tag_len) };
-                    ::pilota::prost::encoding::skip_field(wire_type, tag, buf, ctx)?;
+                    ::pilota::pb::encoding::skip_field(wire_type, tag, buf, ctx)?;
                     let end = buf.chunk().as_ptr();
-                    _unknown_fields.push_back(::pilota::Bytes::copy_from_slice(unsafe {
-                        std::slice::from_raw_parts(begin, end as usize - begin as usize)
-                    }));
+                    let len = end as usize - ctx.raw_bytes_cursor();
+                    let val = ctx.raw_bytes_split_to(len);
+                    _unknown_fields.push_back(val);
                     Ok(())
                 }
             }
@@ -150,21 +127,18 @@ pub mod unknown_fields_pb {
         pub a: ::std::option::Option<A>,
         pub _unknown_fields: ::pilota::BytesVec,
     }
-    impl ::pilota::prost::Message for B {
+    impl ::pilota::pb::Message for B {
         #[inline]
         fn encoded_len(&self) -> usize {
             0 + self.a.as_ref().map_or(0, |msg| {
-                ::pilota::prost::encoding::message::encoded_len(2, msg)
+                ::pilota::pb::encoding::message::encoded_len(2, msg)
             }) + self._unknown_fields.size()
         }
 
         #[allow(unused_variables)]
-        fn encode_raw<B>(&self, buf: &mut B)
-        where
-            B: ::pilota::prost::bytes::BufMut,
-        {
+        fn encode_raw(&self, buf: &mut ::pilota::LinkedBytes) {
             if let Some(_pilota_inner_value) = self.a.as_ref() {
-                ::pilota::prost::encoding::message::encode(2, _pilota_inner_value, buf);
+                ::pilota::pb::encoding::message::encode(2, _pilota_inner_value, buf);
             }
             for bytes in self._unknown_fields.list.iter() {
                 buf.put_slice(bytes.as_ref());
@@ -172,23 +146,20 @@ pub mod unknown_fields_pb {
         }
 
         #[allow(unused_variables)]
-        fn merge_field<B>(
+        fn merge_field(
             &mut self,
             tag: u32,
-            wire_type: ::pilota::prost::encoding::WireType,
-            buf: &mut B,
-            ctx: ::pilota::prost::encoding::DecodeContext,
-        ) -> ::core::result::Result<(), ::pilota::prost::DecodeError>
-        where
-            B: ::pilota::prost::bytes::Buf,
-        {
+            wire_type: ::pilota::pb::encoding::WireType,
+            buf: &mut ::pilota::Bytes,
+            ctx: &mut ::pilota::pb::encoding::DecodeContext,
+        ) -> ::core::result::Result<(), ::pilota::pb::DecodeError> {
             const STRUCT_NAME: &'static str = stringify!(B);
 
             let mut _unknown_fields = &mut self._unknown_fields;
             match tag {
                 2 => {
                     let mut _inner_pilota_value = &mut self.a;
-                    ::pilota::prost::encoding::message::merge(
+                    ::pilota::pb::encoding::message::merge(
                         wire_type,
                         _inner_pilota_value.get_or_insert_with(::core::default::Default::default),
                         buf,
@@ -200,14 +171,11 @@ pub mod unknown_fields_pb {
                     })
                 }
                 _ => {
-                    let tag_value = (tag << 3) | wire_type as u32;
-                    let tag_len = ::pilota::prost::encoding::encoded_len_varint(tag_value as u64);
-                    let begin = unsafe { buf.chunk().as_ptr().sub(tag_len) };
-                    ::pilota::prost::encoding::skip_field(wire_type, tag, buf, ctx)?;
+                    ::pilota::pb::encoding::skip_field(wire_type, tag, buf, ctx)?;
                     let end = buf.chunk().as_ptr();
-                    _unknown_fields.push_back(::pilota::Bytes::copy_from_slice(unsafe {
-                        std::slice::from_raw_parts(begin, end as usize - begin as usize)
-                    }));
+                    let len = end as usize - ctx.raw_bytes_cursor();
+                    let val = ctx.raw_bytes_split_to(len);
+                    _unknown_fields.push_back(val);
                     Ok(())
                 }
             }
@@ -231,21 +199,18 @@ pub mod unknown_fields_pb {
         pub value: ::std::option::Option<::pilota::FastStr>,
         pub _unknown_fields: ::pilota::BytesVec,
     }
-    impl ::pilota::prost::Message for SubMessage {
+    impl ::pilota::pb::Message for SubMessage {
         #[inline]
         fn encoded_len(&self) -> usize {
             0 + self.value.as_ref().map_or(0, |value| {
-                ::pilota::prost::encoding::faststr::encoded_len(2, value)
+                ::pilota::pb::encoding::faststr::encoded_len(2, value)
             }) + self._unknown_fields.size()
         }
 
         #[allow(unused_variables)]
-        fn encode_raw<B>(&self, buf: &mut B)
-        where
-            B: ::pilota::prost::bytes::BufMut,
-        {
+        fn encode_raw(&self, buf: &mut ::pilota::LinkedBytes) {
             if let Some(_pilota_inner_value) = self.value.as_ref() {
-                ::pilota::prost::encoding::faststr::encode(2, _pilota_inner_value, buf);
+                ::pilota::pb::encoding::faststr::encode(2, _pilota_inner_value, buf);
             };
             for bytes in self._unknown_fields.list.iter() {
                 buf.put_slice(bytes.as_ref());
@@ -253,23 +218,20 @@ pub mod unknown_fields_pb {
         }
 
         #[allow(unused_variables)]
-        fn merge_field<B>(
+        fn merge_field(
             &mut self,
             tag: u32,
-            wire_type: ::pilota::prost::encoding::WireType,
-            buf: &mut B,
-            ctx: ::pilota::prost::encoding::DecodeContext,
-        ) -> ::core::result::Result<(), ::pilota::prost::DecodeError>
-        where
-            B: ::pilota::prost::bytes::Buf,
-        {
+            wire_type: ::pilota::pb::encoding::WireType,
+            buf: &mut ::pilota::Bytes,
+            ctx: &mut ::pilota::pb::encoding::DecodeContext,
+        ) -> ::core::result::Result<(), ::pilota::pb::DecodeError> {
             const STRUCT_NAME: &'static str = stringify!(SubMessage);
 
             let mut _unknown_fields = &mut self._unknown_fields;
             match tag {
                 2 => {
                     let mut _inner_pilota_value = &mut self.value;
-                    ::pilota::prost::encoding::faststr::merge(
+                    ::pilota::pb::encoding::faststr::merge(
                         wire_type,
                         _inner_pilota_value.get_or_insert_with(::core::default::Default::default),
                         buf,
@@ -281,14 +243,11 @@ pub mod unknown_fields_pb {
                     })
                 }
                 _ => {
-                    let tag_value = (tag << 3) | wire_type as u32;
-                    let tag_len = ::pilota::prost::encoding::encoded_len_varint(tag_value as u64);
-                    let begin = unsafe { buf.chunk().as_ptr().sub(tag_len) };
-                    ::pilota::prost::encoding::skip_field(wire_type, tag, buf, ctx)?;
+                    ::pilota::pb::encoding::skip_field(wire_type, tag, buf, ctx)?;
                     let end = buf.chunk().as_ptr();
-                    _unknown_fields.push_back(::pilota::Bytes::copy_from_slice(unsafe {
-                        std::slice::from_raw_parts(begin, end as usize - begin as usize)
-                    }));
+                    let len = end as usize - ctx.raw_bytes_cursor();
+                    let val = ctx.raw_bytes_split_to(len);
+                    _unknown_fields.push_back(val);
                     Ok(())
                 }
             }
@@ -314,28 +273,25 @@ pub mod unknown_fields_pb {
         pub sub_messages: ::std::vec::Vec<SubMessage>,
         pub _unknown_fields: ::pilota::BytesVec,
     }
-    impl ::pilota::prost::Message for Message {
+    impl ::pilota::pb::Message for Message {
         #[inline]
         fn encoded_len(&self) -> usize {
-            0 + ::pilota::prost::encoding::faststr::encoded_len(1, &self.uid)
+            0 + ::pilota::pb::encoding::faststr::encoded_len(1, &self.uid)
                 + self.value.as_ref().map_or(0, |value| {
-                    ::pilota::prost::encoding::faststr::encoded_len(2, value)
+                    ::pilota::pb::encoding::faststr::encoded_len(2, value)
                 })
-                + ::pilota::prost::encoding::message::encoded_len_repeated(3, &self.sub_messages)
+                + ::pilota::pb::encoding::message::encoded_len_repeated(3, &self.sub_messages)
                 + self._unknown_fields.size()
         }
 
         #[allow(unused_variables)]
-        fn encode_raw<B>(&self, buf: &mut B)
-        where
-            B: ::pilota::prost::bytes::BufMut,
-        {
-            ::pilota::prost::encoding::faststr::encode(1, &self.uid, buf);
+        fn encode_raw(&self, buf: &mut ::pilota::LinkedBytes) {
+            ::pilota::pb::encoding::faststr::encode(1, &self.uid, buf);
             if let Some(_pilota_inner_value) = self.value.as_ref() {
-                ::pilota::prost::encoding::faststr::encode(2, _pilota_inner_value, buf);
+                ::pilota::pb::encoding::faststr::encode(2, _pilota_inner_value, buf);
             };
             for msg in &self.sub_messages {
-                ::pilota::prost::encoding::message::encode(3, msg, buf);
+                ::pilota::pb::encoding::message::encode(3, msg, buf);
             }
             for bytes in self._unknown_fields.list.iter() {
                 buf.put_slice(bytes.as_ref());
@@ -343,36 +299,28 @@ pub mod unknown_fields_pb {
         }
 
         #[allow(unused_variables)]
-        fn merge_field<B>(
+        fn merge_field(
             &mut self,
             tag: u32,
-            wire_type: ::pilota::prost::encoding::WireType,
-            buf: &mut B,
-            ctx: ::pilota::prost::encoding::DecodeContext,
-        ) -> ::core::result::Result<(), ::pilota::prost::DecodeError>
-        where
-            B: ::pilota::prost::bytes::Buf,
-        {
+            wire_type: ::pilota::pb::encoding::WireType,
+            buf: &mut ::pilota::Bytes,
+            ctx: &mut ::pilota::pb::encoding::DecodeContext,
+        ) -> ::core::result::Result<(), ::pilota::pb::DecodeError> {
             const STRUCT_NAME: &'static str = stringify!(Message);
 
             let mut _unknown_fields = &mut self._unknown_fields;
             match tag {
                 1 => {
                     let mut _inner_pilota_value = &mut self.uid;
-                    ::pilota::prost::encoding::faststr::merge(
-                        wire_type,
-                        _inner_pilota_value,
-                        buf,
-                        ctx,
-                    )
-                    .map_err(|mut error| {
-                        error.push(STRUCT_NAME, stringify!(uid));
-                        error
-                    })
+                    ::pilota::pb::encoding::faststr::merge(wire_type, _inner_pilota_value, buf, ctx)
+                        .map_err(|mut error| {
+                            error.push(STRUCT_NAME, stringify!(uid));
+                            error
+                        })
                 }
                 2 => {
                     let mut _inner_pilota_value = &mut self.value;
-                    ::pilota::prost::encoding::faststr::merge(
+                    ::pilota::pb::encoding::faststr::merge(
                         wire_type,
                         _inner_pilota_value.get_or_insert_with(::core::default::Default::default),
                         buf,
@@ -385,7 +333,7 @@ pub mod unknown_fields_pb {
                 }
                 3 => {
                     let mut _inner_pilota_value = &mut self.sub_messages;
-                    ::pilota::prost::encoding::message::merge_repeated(
+                    ::pilota::pb::encoding::message::merge_repeated(
                         wire_type,
                         _inner_pilota_value,
                         buf,
@@ -397,14 +345,11 @@ pub mod unknown_fields_pb {
                     })
                 }
                 _ => {
-                    let tag_value = (tag << 3) | wire_type as u32;
-                    let tag_len = ::pilota::prost::encoding::encoded_len_varint(tag_value as u64);
-                    let begin = unsafe { buf.chunk().as_ptr().sub(tag_len) };
-                    ::pilota::prost::encoding::skip_field(wire_type, tag, buf, ctx)?;
+                    ::pilota::pb::encoding::skip_field(wire_type, tag, buf, ctx)?;
                     let end = buf.chunk().as_ptr();
-                    _unknown_fields.push_back(::pilota::Bytes::copy_from_slice(unsafe {
-                        std::slice::from_raw_parts(begin, end as usize - begin as usize)
-                    }));
+                    let len = end as usize - ctx.raw_bytes_cursor();
+                    let val = ctx.raw_bytes_split_to(len);
+                    _unknown_fields.push_back(val);
                     Ok(())
                 }
             }
