+++ pilota/src/lib.rs
@@ -4,7 +4,6 @@
 #![cfg_attr(not(doctest), doc = include_str!("../README.md"))]
 
 pub mod pb;
-pub mod prost;
 pub mod thrift;
 
 // reexport
