--- pilota/src/prost/README.md
@@ -1,1 +1,0 @@
-The files in this directory are copied from the [Prost](https://github.com/tokio-rs/prost) project and modified.
